move things properly
